generated_java_autonomous_vehicle_multi_threaded/src/autonomous_vehicle_multi_threaded/SET_ACCTest.java
package autonomous_vehicle_multi_threaded;

import static org.junit.Assert.*;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

public class SET_ACCTest {
	private ref1_kinectic machine;
	
	@Before
	public void setUp() {
		machine = new ref1_kinectic();	
	}

    @After
    public void tearDown() {
            //
    }

	@Test
	public void SET_ACC_01() {
		Integer Vehicle = 11;
		Integer H = 1; // car's height
		Integer W = 1; // car's width
		Integer X = 0; // car's position, x-coordinate
		Integer Y = 0; // car's position, y-coordinate
		Integer Max = 1; // car's maximum velocity
		
		ADD_VEHICLE ac = new ADD_VEHICLE(machine);
		ac.run_ADD_VEHICLE(H, Vehicle, W, X, Y, Max);
		assertTrue(machine.get_vehicles().has(Vehicle));
		
		Integer A = 55;
		SET_ACC sa = new SET_ACC(machine);
		sa.run_SET_ACC(A, Vehicle);

		Integer newA = machine.get_acc().elementImage(Vehicle).choose();
		assertTrue(A==newA);
	}

}
